+++ app/core/security.py
@@ -1,8 +1,8 @@
 from passlib.context import CryptContext
 from jose import jwt, JWTError
 from datetime import datetime, timedelta
-from env_vars import SECRET_KEY, ALGORITHM, ACCESS_TOKEN_EXPIRE_HOURS, OPENAI_API_KEY
-from services.external_request import ExternalRequest
+from app.core.config import SECRET_KEY, ALGORITHM, ACCESS_TOKEN_EXPIRE_HOURS, OPENAI_API_KEY
+from app.services.external_request import ExternalRequest
 
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
@@ -25,23 +25,3 @@ def decode_access_token(token: str):
         return payload
     except JWTError:
         return None
-
-async def get_ai_response(prompt: str) -> str:
-    headers = {
-        "Authorization": f"Bearer {OPENAI_API_KEY}",
-        "Content-Type": "application/json"
-    }
-    data = {
-        "model": "openai/gpt-3.5-turbo",
-        "messages": [{"role": "user", "content": prompt}]
-    }
-    response = await ExternalRequest.send(
-        url="https://openrouter.ai/api/v1/chat/completions",
-        method="POST",
-        headers=headers,
-        json=data
-    )
-    if response is not None:
-        response_json = response.json()
-        return response_json.get("choices", [])[0].get("message", {}).get("content", "").strip()
-    return "[Error: No response from AI service]"
